major rewrite of design principles
The original design principles were:
* Cohesion
* Consistency
* Endomorphism
* Fluency

That is now:
* Cognitive consistency and fluency
* Cohesion
* Expressiveness

The Cohesion section is mostly the same, except in goes into more
detail about the advantages and disadvantages of the high degree of
cohesion pursued by dplyr. Bioconductor is already very cohesive (it
is a goal of all software engineers). dplyr takes API cohesion to the
level that individual function calls are effectively coupled (like
group_by->summarize always being used together). That is unusual and
worth discussing in my opinion.

The Consistency section seemed to be mixing some things. The notion of
being consistent with the user's intuition about strand got moved to
the 'Cognitive consistency' section. The other part seemed mostly
related to expressiveness, i.e., names indicating intent, so that got
moved to the Expressiveness section.

Endomorphism is a good design principle, although it seemed to mostly
flow from our decision to use a single data structure. I mentioned it
in the "Data model" section, but I'm not sure it was really a core
principle, since there was no tension, i.e., no way that we could make
a non-endomorphic API given that everything is a GRanges. So this
section got dropped.

Fluency seemed to summarize some of the other points and then talked
about the intuitive chaining syntax. Since this relates to how the
user thinks about the problem (natural language, pipelines), I worked
this into the cognitive consistency section. It seems to work.

I moved the Opportunities section to Discussion, which is more
conventional for these types of papers. But we need to develop that
section more, or at least better integrate what is there now.